refactor: share stateless responses backend
