+++ packages/agent-openrouter/agent-openrouter.cabal
@@ -66,7 +66,6 @@ test-suite agent-openrouter-test
                     , Agent.OpenRouter.CredentialSpec
                     , Agent.OpenRouter.ErrorSpec
                     , Agent.OpenRouter.FunctionalSpec
-                    , Agent.OpenRouter.LoopBackendSpec
                     , Agent.OpenRouter.RequestSpec
                     , Agent.OpenRouter.UsageSpec
     ghc-options:      -threaded
