+++ packages/agent-openrouter/src/Agent/OpenRouter/LoopBackend.hs
@@ -11,13 +11,8 @@ module Agent.OpenRouter.LoopBackend
     ) where
 
 import Agent.Error (ApiError)
-import Agent.Loop (Backend(..), LoopEvent, TurnInput, TurnOutput)
-import Agent.OpenAI.LoopBackend
-    ( responseToTurnOutput
-    , streamEventToLoopEvent
-    , turnInputsToItems
-    , withRequestInput
-    )
+import Agent.Loop (Backend)
+import Agent.OpenAI.LoopBackend (statelessResponsesBackend)
 import Agent.OpenAI.Responses.Types
 import Agent.Provider
     ( TokenProvider
@@ -51,28 +46,4 @@ openRouterBackendWith
     -> IO ResponseCreateParams
     -> IORef [ResponseItem]
     -> Backend
-openRouterBackendWith send getParams transcript = Backend \_previousResponseId inputs onEvent -> do
-    baseParams <- getParams
-    submitOpenRouterTurn send baseParams transcript inputs onEvent
-
-submitOpenRouterTurn
-    :: (ResponseCreateParams
-        -> (ResponseStreamEvent -> IO ())
-        -> IO (Either ApiError Response))
-    -> ResponseCreateParams
-    -> IORef [ResponseItem]
-    -> [TurnInput]
-    -> (LoopEvent -> IO ())
-    -> IO (Either ApiError TurnOutput)
-submitOpenRouterTurn send baseParams transcript inputs onEvent = do
-    history <- readIORef transcript
-    let newItems = turnInputsToItems inputs
-        requestItems = history <> newItems
-        request = withRequestInput baseParams requestItems
-    result <- send request \event ->
-        mapM_ onEvent (streamEventToLoopEvent event)
-    case result of
-        Left err -> pure (Left err)
-        Right response -> do
-            writeIORef transcript (requestItems <> response.output)
-            pure (Right (responseToTurnOutput response))
+openRouterBackendWith = statelessResponsesBackend
