+++ packages/agent-openrouter/test/Spec.hs
@@ -6,7 +6,6 @@ import qualified Agent.OpenRouter.ClientSpec as ClientSpec
 import qualified Agent.OpenRouter.CredentialSpec as CredentialSpec
 import qualified Agent.OpenRouter.ErrorSpec as ErrorSpec
 import qualified Agent.OpenRouter.FunctionalSpec as FunctionalSpec
-import qualified Agent.OpenRouter.LoopBackendSpec as LoopBackendSpec
 import qualified Agent.OpenRouter.RequestSpec as RequestSpec
 import qualified Agent.OpenRouter.UsageSpec as UsageSpec
 
@@ -16,6 +15,5 @@ main = hspec do
     ErrorSpec.spec
     CredentialSpec.spec
     ClientSpec.spec
-    LoopBackendSpec.spec
     FunctionalSpec.spec
     UsageSpec.spec
