+++ packages/agent-xai/agent-xai.cabal
@@ -63,7 +63,6 @@ test-suite agent-xai-test
     other-modules:    Agent.XAI.AuthSpec
                     , Agent.XAI.ClientSpec
                     , Agent.XAI.FunctionalSpec
-                    , Agent.XAI.LoopBackendSpec
                     , Agent.XAI.ResponsesSpec
                     , Agent.XAI.UsageSpec
     ghc-options:      -threaded
